refactor: got rid of unused authorizers
