+++ internal/authorizer/internal.go
@@ -5,34 +5,34 @@ import (
 	"github.com/Codename-Uranium/tunnel/pkg/xerror"
 )
 
-type InternalAuthorizer struct {
+type JWTAuthorizer struct {
 	checker *xcrypto.JWTChecker
 	running bool
 }
 
-func NewInternalAuthorizer(keyKeeper xcrypto.KeyStore) (*InternalAuthorizer, error) {
+func NewJWT(keyKeeper xcrypto.KeyStore) (*JWTAuthorizer, error) {
 	checker, err := xcrypto.NewJWTChecker(keyKeeper)
 
 	if err != nil {
 		return nil, err
 	}
 
-	return &InternalAuthorizer{
+	return &JWTAuthorizer{
 		checker: checker,
 		running: true,
 	}, nil
 }
 
-func (d *InternalAuthorizer) Shutdown() error {
+func (d *JWTAuthorizer) Shutdown() error {
 	d.running = false
 	return nil
 }
 
-func (d *InternalAuthorizer) Running() bool {
+func (d *JWTAuthorizer) Running() bool {
 	return d.running
 }
 
-func (d *InternalAuthorizer) Authenticate(tokenString string, myAudience string) (*xcrypto.ClientClaims, error) {
+func (d *JWTAuthorizer) Authenticate(tokenString string, myAudience string) (*xcrypto.ClientClaims, error) {
 	var claims xcrypto.ClientClaims
 
 	err := d.checker.Parse(tokenString, &claims)
